requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.4.0
celery==4.4.7
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
configparser==5.0.1
cryptography==3.4.7
DateTime==4.3
defusedxml==0.7.1
Django==3.1.2
django-allauth==0.44.0
django-pipeline==2.0.6
django-redis==4.12.1
et-xmlfile==1.0.1
flower==0.9.7
humanize==3.5.0
idna==2.10
jdcal==1.4.1
kombu==4.6.11
oauthlib==3.1.0
Pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.18
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytils==0.3
pytz==2020.1
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
schedule==1.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.1
sqlparse==0.4.1
tornado==6.1
urllib3==1.26.2
vine==1.3.0
wcwidth==0.2.5
zope.interface==5.2.0